JpaRepository<Role, Integer> has been changed to JpaRepository<Role, Long>

+++ Management-App/src/main/java/pl/dahdev/managementapp/repository/ActivationRepository.java
@@ -5,7 +5,7 @@ import pl.dahdev.managementapp.model.Activation;
 
 import java.util.Optional;
 
-public interface ActivationRepository extends JpaRepository<Activation, Integer> {
+public interface ActivationRepository extends JpaRepository<Activation, Long> {
 
     Optional<Activation> findById(long id);
